add mapping exoplanet to user

+++ spacedust-api/api/exoplanets/query/index.js
@@ -1,34 +1,35 @@
 const Exoplanet = require('./../model');
 
-const getExoplanets = async () => {
+const getExoplanets = async owner => {
     try {
-        return await Exoplanet.find({});
+        return await Exoplanet.find({ owner });
     } catch (e) {
         return err;
 
     }
 };
 
-const getExoplanet = async id => {
+const getExoplanet = async (owner, id) => {
     try {
-        return await Exoplanet.findOne({_id: id});
+        return await Exoplanet.findOne({ owner, _id: id });
     } catch (e) {
         return e;
     }
 };
 
-const createExoplanet = async exoplanetData => {
+const createExoplanet = async (owner, exoplanetData) => {
     try {
         const newExoplanet = new Exoplanet(exoplanetData);
+        newExoplanet.owner = owner;
         return await newExoplanet.save();
     } catch (e) {
         return e;
     }
 };
 
-const deleteExoplanet = async id => {
+const deleteExoplanet = async (owner, id) => {
     try {
-        return await Exoplanet.findOneAndRemove({_id: id});
+        return await Exoplanet.findOneAndRemove({ owner, _id: id});
     } catch (e) {
         return e;
     }
